% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{Formula_Data}
\alias{Formula_Data}
\alias{Formula_Names}
\title{Formula_Data}
\usage{
Formula_Data(x, data, subset, na.action = na.pass, ...)

Formula_Names(x, data = NULL, ...)
}
\arguments{
\item{x}{Formel ~a+b+c, a+b+c~d}

\item{data}{Daten als data.frame}

\item{subset, na.action}{geht nicht default ist na.pass}
}
\value{
Formula_Data: list mit "X_data","Y_data","xname","yname",
"Z_data",  "zname", "formula", "condition", "formula.orginal", "digits", "type"
}
\description{
Alte Funktion von \code{errate_statistik2(...)} \code{bzw APA2(...)}

Formula_Names interne Funktion in Formula_Data() und in model_info.formula verwendet
}
