% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-helper.R, R/makeFormula.R, R/misc.R
\name{formula_helper}
\alias{formula_helper}
\alias{clean_dots_formula}
\alias{formula_split}
\alias{make_formula}
\alias{check_data}
\alias{makeNamesNum}
\title{Formel-Objekte bearbeiten}
\usage{
formula_helper(...)

clean_dots_formula(x, data = NULL, names_data = names(data))

formula_split(x)

make_formula(measurevar, groupvars)

check_data(data, vars = NULL)

makeNamesNum(measure, data, meAsNum = grepl("^[[:digit:]]", measure))
}
\arguments{
\item{...}{weitere einstellungen}

\item{x}{Formula}

\item{data, names_data}{data.frame}

\item{measurevar, groupvars}{mamen als strings}

\item{vars}{in check_data}

\item{measure}{Variablen}

\item{meAsNum}{logical welche sind Zahlen}

\item{data}{Daten als data.frame}
}
\value{
liste

clean_dots_formula: formula - Objekt

formula_split: liste  formula, condition, facet_type

check_data: Logical und wen FALSE ueber cat die  Objekt die falsch sind

string( )
}
\description{
Formulas

clean_dots_formula: Formel bereinigen

formula_split stolen from mosaic ggformula

make_formula: Formel erstellen in \code{berechne_all(...)} verwendet. 
Hier wird \code{cbind(a,b,c)~g} ausgegebeb.

check_data: prueft ob variablen vorhanden sind bzw ob Faelle NA sind.

makeNamesNum: aus Nummern die Namen extrahieren
}
\examples{

data <- data.frame(x = NA, y = NA, z = NA)
stp25formula:::clean_dots_formula(x ~ y, data)
stp25formula:::clean_dots_formula(. ~ x + y, data)
stp25formula:::clean_dots_formula(x + y ~ ., data)
stp25formula:::clean_dots_formula(~., data)
stp25formula:::formula_split(a+b~x|y)



make_formula("a", "b")
make_formula("a", c("b","c"))
make_formula("a", ~b+c)
make_formula(c("a", "d"), c("b","c"))


dat<- fata.frame(sex=1:2, m1=1:2, m3=1:2)
check_data(dat, c("m1", "m2") )
check_data(dat, c("m1", "sex") )


  measure <- c("geschl", "1" , "3:5", 1)
  stp25formula:::makeNamesNum(measure,  data=dat)

}
