% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data3.R
\name{prepare_data2.data.frame}
\alias{prepare_data2.data.frame}
\title{prepare_data2.data.frame}
\usage{
\method{prepare_data2}{data.frame}(data, ..., by = "1", groups = NULL,
  na.action = na.pass)
}
\arguments{
\item{...}{Namen oder Nummern (y-Variablen))}

\item{by}{x-Variablen}
}
\description{
prepare_data2.data.frame
}
