% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data3.R
\name{prepare_data2.formula}
\alias{prepare_data2.formula}
\title{prepare_data2.formula}
\usage{
\method{prepare_data2}{formula}(x, data, groups = NULL,
  na.action = na.pass)
}
\arguments{
\item{x}{formel}

\item{data}{data.frame}

\item{groups}{condition}

\item{na.action}{na.pass, na.omit}
}
\description{
prepare_data2.formula
}
