% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data2.R
\name{prepare_data2}
\alias{prepare_data2}
\alias{prepare_data2.formula}
\alias{prepare_data2.data.frame}
\alias{prepare_data2.NULL}
\title{prepare_data2}
\usage{
prepare_data2(...)

\method{prepare_data2}{formula}(x, data, subset, na.action = na.pass,
  groups = NULL)

\method{prepare_data2}{data.frame}(data, ..., by = "1", groups = NULL,
  subset, na.action = na.pass)

\method{prepare_data2}{NULL}()
}
\arguments{
\item{...}{Variablen}

\item{x}{formel}

\item{data}{daten}

\item{subset}{nicht implementiert}

\item{na.action}{nicht zum aendern}

\item{groups}{Block codition}

\item{by}{Gruppe}

\item{groups}{Block codition}
}
\value{
Liste mit Namen und Daten
data,
measure.vars ,
group.vars,
condition.vars,
formula,
by,
measure,
row_name,
col_name,
measure.class,
group.class,
condition.class,
digits
}
\description{
Funktion wird zum Aufbereiten der Daten verwendet. Die Daten werden als
 tibble::as_tibble() weitergegeben.
}
\details{
Tabelle.default()
 errate_statistik3()
 APA.formula()
 corr_plot.formula()
 Hmisc_rcorr() also APA_Correlation
}
\examples{

dat<- data.frame(sex=1:2, m1=1:2,m2=1:2, m3=1:2, m4=1:2, m5=1:2, m6=1:2, geschl=1:2)

prepare_data2(~ m1 + m2 + m3 + m4, dat)
prepare_data2(~ log(m1) + m2 + m3 + m4, dat)
prepare_data2(~ m1[1] + m2 + m3 + m4, dat)
prepare_data2(~ m1[1] + m2 + m3[4,median] + m4, dat)

prepare_data2(dat, m1, m2, m3, m4)
prepare_data2(dat, 4:7)
prepare_data2(dat, m1[1], m2, m3, m4)
prepare_data2(dat, m1[1], m2, m3[4,median], m4)
prepare_data2(dat, m1 , m2, m3=median, m4)
prepare_data2(dat, m1, m2, m3 , m4, by =  ~ geschl)
prepare_data2(dat, m1[4, median], m2, m3 , m4[5], by =  ~ geschl)


}
