% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz.R
\docType{package}
\name{stp25formula-package}
\alias{stp25formula}
\alias{stp25formula-package}
\title{stp25formula}
\description{
Auflosen von Formeln und Daten.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stp4/stp25formula}
}

}
\author{
\strong{Maintainer}: Wolfgang Peter \email{w.peter@tirol.com}

}
